v0.6.7: bump version + cut release section + close umbrella
Release mechanics:
- package.json 0.6.6 → 0.6.7 so provenance.serverVersion matches the
  shipped tag (caught by the smoke-test agent — the live server was
  reporting 0.6.6 against an `[Unreleased]` body of 0.6.7 features).
- CHANGELOG.md: cut `[Unreleased]` → `[0.6.7] - 2026-05-04` with a
  fresh empty `[Unreleased]` above for post-release work.

Issue #13 — modular layering, closed with evidence:
- Split src/tools/file-tools.ts (592 lines) into a 29-line barrel
  plus four focused sub-modules under src/tools/file/:
    file/shared.ts            — pagination + sandboxed-walk helpers
    file/read-module.ts       — agda_read_module
    file/list-modules.ts      — agda_list_modules
    file/check-postulates.ts  — agda_check_postulates
                                (with findPostulates extracted as a
                                pure helper for direct unit testing)
    file/search-definitions.ts — agda_search_definitions
- Every src/ file now ≤ 500 lines (largest is goal-tools.ts at 494).
- ARCHITECTURE.md: clarified that the layering diagram describes
  data flow, not module imports. The audit's "reverse import"
  findings (session/register-* importing tool-helpers, etc.) are
  explicitly identified as cross-cutting infrastructure rather than
  data-flow violations. The constraint is now stated as: domain
  logic stays in its layer; shared types/registration plumbing may
  cross.

Fixture maintenance (smoke-test follow-up):
- AbstractHoleMultiple.agda, NestedAbstractHole.agda,
  MixedVisibleInvisible.agda all carried docstrings claiming their
  abstract-block holes are reported as invisible goals (NamedMeta).
  On Agda 2.8.0 they come back as visible goals — the abstract
  keyword no longer suppresses interaction-layer tracking.
  Classification (`ok-with-holes`) is correct under both reportings
  via the merged source-hole + protocol-counts signal. Updated each
  docstring to reflect 2.8.0 reality and pointed readers at
  InferredMeta.agda (uses `_`-underscore inferred metas) for the
  invisible-only-goal fixture that still exercises the merged path
  on 2.8.0 — asserted at agda-load.test.ts:164–170.

Issue closures (issues #9, #12, #13, #14 — all v0.6.7-tracked):
- Audited each issue's stated acceptance criteria against the live
  code, posted file:line evidence in the closure comment, and
  closed. Umbrella #14 closed as superseded — all nine sub-issues
  are now done.

Verified: full test suite passes (1288 / 0), TypeScript build clean.
