+++ src/tools/file-tools.ts
@@ -1,592 +1,29 @@
 // MIT License — see LICENSE
 //
-// File-based navigation tools: reading modules, listing modules,
-// checking postulates, searching definitions (pure filesystem, no Agda session)
+// Barrel registration for the file-navigation tool family
+// (`agda_read_module`, `agda_list_modules`, `agda_check_postulates`,
+// `agda_search_definitions`). Pure filesystem tools — none of these
+// require an active Agda session beyond best-effort version detection
+// for the source-extension filter. The actual callbacks live in
+// focused per-tool modules under `src/tools/file/` so this barrel
+// stays under the 500-line ceiling and each tool's logic can be read
+// in isolation.
 
 import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import { z } from "zod";
-import { resolve, relative, join } from "node:path";
-import { readFileSync, existsSync, readdirSync } from "node:fs";
+
 import type { AgdaSession } from "../agda-process.js";
-import { isAgdaSourceFile, filePathDescription } from "../agda/version-support.js";
-import { matchesTypePattern } from "../agda/agent-ux.js";
-import { logger } from "../agda/logger.js";
-import { PathSandboxError, resolveExistingPathWithinRoot, resolveFileWithinRoot } from "../repo-root.js";
-import { missingPathToolError, ToolInvocationError, registerTextTool } from "./tool-helpers.js";
-import { extractLiterateCode } from "../session/literate-extraction.js";
-
-/**
- * Default page size for `agda_list_modules`. Sized so a single response
- * comfortably fits inside an MCP client's per-tool token budget on a
- * many-hundred-module project — see §2.4 of the agent UX bug report.
- */
-const LIST_MODULES_DEFAULT_LIMIT = 25;
-/** Hard cap so a caller can't ask for, say, 100k results in one shot. */
-const LIST_MODULES_MAX_LIMIT = 500;
-
-function relativeToRequestedRoot(repoRoot: string, requestedRoot: string, relativePath = ""): string {
-  const requestedBase = relative(repoRoot, requestedRoot);
-  return relativePath ? join(requestedBase, relativePath) : requestedBase;
-}
-
-function resolveExistingChildWithinRoot(repoRoot: string, path: string): string | null {
-  try {
-    return resolveExistingPathWithinRoot(repoRoot, path);
-  } catch (error) {
-    if (error instanceof PathSandboxError) {
-      return null;
-    }
-    throw error;
-  }
-}
+import { register as registerCheckPostulates } from "./file/check-postulates.js";
+import { register as registerListModules } from "./file/list-modules.js";
+import { register as registerReadModule } from "./file/read-module.js";
+import { register as registerSearchDefinitions } from "./file/search-definitions.js";
 
 export function register(
   server: McpServer,
   session: AgdaSession,
   repoRoot: string,
 ): void {
-  registerTextTool({
-    server,
-    name: "agda_read_module",
-    description: "Read an Agda module file and return its contents with line numbers. For literate files (.lagda.md, .lagda.tex, etc.), use `codeOnly: true` to extract just the Agda code blocks, stripping prose wrappers.",
-    category: "navigation",
-    inputSchema: {
-      file: z.string().describe(filePathDescription(session.getAgdaVersion() ?? undefined)),
-      codeOnly: z.boolean().optional().describe("When true, extract only Agda code blocks from literate files, stripping prose. Has no effect on plain .agda files."),
-    },
-    outputDataSchema: z.object({
-      text: z.string(),
-      file: z.string(),
-      lineCount: z.number(),
-      literateFormat: z.string().nullable(),
-      codeOnly: z.boolean(),
-      codeBlockCount: z.number().nullable(),
-    }),
-    callback: async ({ file, codeOnly }: { file: string; codeOnly?: boolean }) => {
-      const requestedFilePath = resolveFileWithinRoot(repoRoot, file);
-      if (!existsSync(requestedFilePath)) {
-        throw missingPathToolError("file", requestedFilePath);
-      }
-      const filePath = resolveExistingPathWithinRoot(repoRoot, requestedFilePath);
-      const fileContent = readFileSync(filePath, "utf-8");
-      const canonicalRoot = resolveExistingPathWithinRoot(repoRoot, ".");
-      const relPath = relative(canonicalRoot, filePath);
-      const lineCount = fileContent.split("\n").length;
-      const isCodeOnly = Boolean(codeOnly);
-
-      if (isCodeOnly) {
-        const extraction = extractLiterateCode(filePath, fileContent);
-        if (extraction.format === null) {
-          // Plain .agda — codeOnly has no effect, return as normal
-          const numbered = fileContent
-            .split("\n")
-            .map((line, i) => `${String(i + 1).padStart(4)} | ${line}`)
-            .join("\n");
-          return {
-            text: `## ${relPath}\n\n\`\`\`agda\n${numbered}\n\`\`\``,
-            data: {
-              file: relPath,
-              lineCount,
-              literateFormat: null,
-              codeOnly: true,
-              codeBlockCount: null,
-            },
-          };
-        }
-
-        if (extraction.blocks.length === 0) {
-          return {
-            text: `## ${relPath} (code only)\n\nNo Agda code blocks found in this ${extraction.format} literate file.`,
-            data: {
-              file: relPath,
-              lineCount,
-              literateFormat: extraction.format,
-              codeOnly: true,
-              codeBlockCount: 0,
-            },
-          };
-        }
-
-        let output = `## ${relPath} (code only — ${extraction.format} format, ${extraction.blocks.length} block(s))\n\n`;
-        for (const block of extraction.blocks) {
-          const numbered = block.code
-            .split("\n")
-            .map((line, lineIdx) => `${String(block.startLine + lineIdx).padStart(4)} | ${line}`)
-            .join("\n");
-          output += `\`\`\`agda\n${numbered}\n\`\`\`\n\n`;
-        }
-        return {
-          text: output.trimEnd(),
-          data: {
-            file: relPath,
-            lineCount,
-            literateFormat: extraction.format,
-            codeOnly: true,
-            codeBlockCount: extraction.blocks.length,
-          },
-        };
-      }
-
-      const numbered = fileContent
-        .split("\n")
-        .map((line, i) => `${String(i + 1).padStart(4)} | ${line}`)
-        .join("\n");
-      return {
-        text: `## ${relPath}\n\n\`\`\`agda\n${numbered}\n\`\`\``,
-        data: {
-          file: relPath,
-          lineCount,
-          literateFormat: null,
-          codeOnly: false,
-          codeBlockCount: null,
-        },
-      };
-    },
-  });
-
-  registerTextTool({
-    server,
-    name: "agda_list_modules",
-    description: "List Agda modules in a directory tier (MathLib, Foundation, Kernel, Research, Extensions, etc.). Always reports the total module count; paginated to keep responses small (default page size 25). Use `offset` to scroll, `limit` to enlarge the page, and `pattern` for a case-insensitive substring filter on the relative path.",
-    category: "navigation",
-    inputSchema: {
-      tier: z.string().describe("The tier to list, e.g. 'Kernel', 'Foundation', 'MathLib'"),
-      offset: z.number().int().min(0).optional().describe("0-based starting index into the sorted result list. Defaults to 0."),
-      limit: z.number().int().min(1).max(LIST_MODULES_MAX_LIMIT).optional().describe(
-        `Maximum number of modules to return in this page. Defaults to ${LIST_MODULES_DEFAULT_LIMIT}; capped at ${LIST_MODULES_MAX_LIMIT}.`,
-      ),
-      pattern: z.string().optional().describe("Case-insensitive substring filter applied to each module's relative path before pagination."),
-    },
-    outputDataSchema: z.object({
-      text: z.string(),
-      tier: z.string(),
-      pattern: z.string().nullable(),
-      total: z.number(),
-      filtered: z.number(),
-      offset: z.number(),
-      limit: z.number(),
-      modules: z.array(z.string()),
-      hasMore: z.boolean(),
-      nextOffset: z.number().nullable(),
-      unreadableSubtrees: z.array(z.string()),
-    }),
-    callback: async ({ tier, offset, limit, pattern }: { tier: string; offset?: number; limit?: number; pattern?: string }) => {
-      const requestedTierDir = resolveFileWithinRoot(repoRoot, join("agda", tier));
-      if (!existsSync(requestedTierDir)) {
-        throw new ToolInvocationError({
-          message: `Tier directory not found: agda/${tier}`,
-          classification: "not-found",
-          diagnostics: [
-            {
-              severity: "error",
-              message: `Tier directory not found: agda/${tier}`,
-              code: "not-found",
-              nextAction:
-                "Pass an existing top-level directory under agda/ (e.g. MathLib, Foundation, Kernel, Research, Extensions, TrustedCompute). " +
-                "Use `agda_file_list` with no `tier` to see the available top-level directories in this project.",
-            },
-            {
-              severity: "info",
-              message: "Available tiers: MathLib, Foundation, Kernel, Research, Extensions, TrustedCompute",
-              code: "available-tiers",
-            },
-          ],
-          data: { tier },
-          text: `Tier directory not found: agda/${tier}\nAvailable: MathLib, Foundation, Kernel, Research, Extensions, TrustedCompute`,
-        });
-      }
-      const tierDir = resolveExistingPathWithinRoot(repoRoot, requestedTierDir);
-      // Ensure version detection has run so isAgdaSourceFile() filters by the
-      // actually-installed Agda's supported extensions rather than being
-      // permissive (all extensions) when this tool is invoked first.
-      if (!session.getAgdaVersion()) {
-        try { await session.query.showVersion(); } catch (err) {
-          logger.trace("agda_list_modules: version detection best-effort failed", { err });
-        }
-      }
-      const modules: string[] = [];
-      const unreadableDirs: string[] = [];
-      function walk(dir: string, requestedDir: string, displayPrefix: string): void {
-        let entries: import("node:fs").Dirent[];
-        try {
-          entries = readdirSync(dir, { withFileTypes: true });
-        } catch (err) {
-          // One unreadable subdir (permission denied, broken
-          // symlink, ENOTDIR race from a concurrent rename) must
-          // not crash the entire listing. Collect the path so the
-          // response can tell the caller which subtrees were
-          // skipped, then keep walking the siblings.
-          unreadableDirs.push(displayPrefix);
-          logger.trace("agda_list_modules: readdir failed, skipping subtree", { dir, err });
-          return;
-        }
-        for (const entry of entries) {
-          const nextRequestedPath = resolve(requestedDir, entry.name);
-          const nextDisplayPath = join(displayPrefix, entry.name);
-          if (entry.isDirectory()) walk(resolve(dir, entry.name), nextRequestedPath, nextDisplayPath);
-          else if (isAgdaSourceFile(entry.name, session.getAgdaVersion() ?? undefined)) {
-            const canonicalPath = resolveExistingChildWithinRoot(repoRoot, nextRequestedPath);
-            if (canonicalPath) {
-              modules.push(nextDisplayPath);
-            }
-          }
-        }
-      }
-      walk(
-        tierDir,
-        requestedTierDir,
-        relativeToRequestedRoot(repoRoot, requestedTierDir),
-      );
-      modules.sort();
-
-      const normalizedPattern = pattern?.toLowerCase() ?? null;
-      const filtered = normalizedPattern === null
-        ? modules
-        : modules.filter((m) => m.toLowerCase().includes(normalizedPattern));
-
-      const effectiveOffset = offset ?? 0;
-      const effectiveLimit = limit ?? LIST_MODULES_DEFAULT_LIMIT;
-      const page = filtered.slice(effectiveOffset, effectiveOffset + effectiveLimit);
-      const nextOffset = effectiveOffset + page.length;
-      const hasMore = nextOffset < filtered.length;
-
-      const totalLine = normalizedPattern === null
-        ? `**Total:** ${modules.length} modules in \`agda/${tier}\`.`
-        : `**Total:** ${filtered.length} matches for \`${pattern}\` (out of ${modules.length} modules in \`agda/${tier}\`).`;
-      const rangeStart = page.length === 0 ? effectiveOffset : effectiveOffset + 1;
-      const rangeEnd = effectiveOffset + page.length;
-      const showingLine = page.length === 0
-        ? `**Showing:** none — \`offset: ${effectiveOffset}\` is past the end (${filtered.length} total).`
-        : `**Showing:** ${rangeStart}–${rangeEnd} of ${filtered.length}.`;
-
-      const lines = [
-        `## agda/${tier}`,
-        "",
-        totalLine,
-        showingLine,
-        "",
-      ];
-      if (page.length > 0) {
-        for (const m of page) lines.push(`- ${m}`);
-      }
-      if (hasMore) {
-        lines.push("");
-        lines.push(`**More results available.** Re-call with \`offset: ${nextOffset}\` to fetch the next page.`);
-      }
-      if (unreadableDirs.length > 0) {
-        lines.push("");
-        lines.push(`**Skipped ${unreadableDirs.length} unreadable subtree(s):** ${unreadableDirs.join(", ")}. Check file permissions or broken symlinks; modules beneath these directories are not included in the total count.`);
-      }
-      return {
-        text: lines.join("\n"),
-        data: {
-          tier,
-          pattern: pattern ?? null,
-          total: modules.length,
-          filtered: filtered.length,
-          offset: effectiveOffset,
-          limit: effectiveLimit,
-          modules: page,
-          hasMore,
-          nextOffset: hasMore ? nextOffset : null,
-          unreadableSubtrees: unreadableDirs,
-        },
-      };
-    },
-  });
-
-  registerTextTool({
-    server,
-    name: "agda_check_postulates",
-    description: "Check an Agda file for postulate declarations. In Kernel/ files, postulates are forbidden by construction.",
-    category: "navigation",
-    inputSchema: { file: z.string().describe(filePathDescription(session.getAgdaVersion() ?? undefined)) },
-    outputDataSchema: z.object({
-      text: z.string(),
-      file: z.string(),
-      blockCount: z.number(),
-      declarationCount: z.number(),
-      isKernel: z.boolean(),
-      kernelViolation: z.boolean(),
-      blocks: z.array(z.object({
-        line: z.number(),
-        declarations: z.array(z.string()),
-      })),
-    }),
-    callback: async ({ file }: { file: string }) => {
-      const requestedFilePath = resolveFileWithinRoot(repoRoot, file);
-      if (!existsSync(requestedFilePath)) {
-        throw missingPathToolError("file", requestedFilePath);
-      }
-      const filePath = resolveExistingPathWithinRoot(repoRoot, requestedFilePath);
-      const fileContent = readFileSync(filePath, "utf-8");
-      const lines = fileContent.split("\n");
-
-      interface PostulateBlock {
-        /** 1-based line number of the `postulate` keyword */
-        line: number;
-        /** Identifier names declared in the block (may be empty for inline postulate) */
-        declarations: string[];
-      }
-
-      const blocks: PostulateBlock[] = [];
-
-      for (let i = 0; i < lines.length; i++) {
-        const trimmed = lines[i].trim();
-        if (!/^postulate\b/.test(trimmed) && !/^\s+postulate\b/.test(lines[i])) {
-          continue;
-        }
-
-        // Check whether the postulate keyword has an inline declaration on the
-        // same line, e.g. `postulate ax : Set` or `postulate p q : Set`.
-        // If there is content after the keyword, parse it; otherwise collect
-        // the indented lines that follow as the block body.
-        const inlineMatch = /^postulate\s+(\S.*)$/.exec(trimmed);
-        if (inlineMatch) {
-          // Inline form: split on `:` and tokenise LHS to get all names
-          // (handles `postulate p q : Set` → ["p", "q"])
-          const lhs = inlineMatch[1].split(":")[0].trim();
-          const declNames = lhs.split(/\s+/u).filter((t) => t.length > 0 && !t.startsWith("--"));
-          blocks.push({ line: i + 1, declarations: declNames });
-          continue;
-        }
-
-        // Block form: collect following indented lines until de-indentation
-        const keywordIndent = lines[i].match(/^(\s*)/)?.[1].length ?? 0;
-        const declarations: string[] = [];
-        let j = i + 1;
-        while (j < lines.length) {
-          const bodyLine = lines[j];
-          const bodyTrimmed = bodyLine.trim();
-          // Skip blank lines and comment-only lines within the block
-          if (bodyTrimmed === "" || bodyTrimmed.startsWith("--")) {
-            j++;
-            continue;
-          }
-          // A non-blank line that is NOT more indented than the postulate keyword
-          // means we've left the block
-          const bodyIndent = bodyLine.match(/^(\s*)/)?.[1].length ?? 0;
-          if (bodyIndent <= keywordIndent) {
-            break;
-          }
-          // Split on `:` and tokenise LHS — handles `p q : Set` (multiple names)
-          const lhs = bodyTrimmed.split(":")[0].trim();
-          const names = lhs.split(/\s+/u).filter((t) => t.length > 0 && !t.startsWith("--"));
-          declarations.push(...names);
-          j++;
-        }
-        blocks.push({ line: i + 1, declarations });
-      }
-
-      const canonicalRoot = resolveExistingPathWithinRoot(repoRoot, ".");
-      const relPath = relative(canonicalRoot, filePath);
-      const isKernel = relPath.startsWith("agda/Kernel/");
-      let output = `## Postulate check: ${relPath}\n\n`;
-      const totalDecls = blocks.reduce((sum, b) => sum + b.declarations.length, 0);
-      if (blocks.length === 0) {
-        output += "No postulates found. Fully constructive.\n";
-      } else {
-        const declSummary = totalDecls > 0
-          ? ` (${totalDecls} identifier${totalDecls === 1 ? "" : "s"} declared)`
-          : "";
-        output += `**${blocks.length} postulate block${blocks.length === 1 ? "" : "s"} found${declSummary}**`;
-        if (isKernel) output += ` **VIOLATION: postulates are forbidden in Kernel/**`;
-        output += "\n\n";
-        for (const block of blocks) {
-          if (block.declarations.length > 0) {
-            output += `- Line ${block.line}: postulate — ${block.declarations.join(", ")}\n`;
-          } else {
-            output += `- Line ${block.line}: postulate\n`;
-          }
-        }
-      }
-      return {
-        text: output,
-        data: {
-          file: relPath,
-          blockCount: blocks.length,
-          declarationCount: totalDecls,
-          isKernel,
-          kernelViolation: isKernel && blocks.length > 0,
-          blocks: blocks.map((b) => ({
-            line: b.line,
-            declarations: b.declarations,
-          })),
-        },
-      };
-    },
-  });
-
-  registerTextTool({
-    server,
-    name: "agda_search_definitions",
-    description: "Search for a definition, theorem, or type name across Agda modules.",
-    category: "navigation",
-    inputSchema: {
-      query: z.string().optional().describe("The name or pattern to search for"),
-      typePattern: z.string().optional().describe("Type-shape query (wildcard `_` supported), e.g. `_ ≤ _ + _`"),
-      tier: z.string().optional().describe("Optional tier to limit search (Kernel, Foundation, etc.)"),
-    },
-    outputDataSchema: z.object({
-      text: z.string(),
-      mode: z.enum(["name", "type-pattern"]),
-      query: z.string(),
-      tier: z.string().nullable(),
-      matchCount: z.number(),
-      shown: z.number(),
-      truncated: z.boolean(),
-      matches: z.array(z.object({
-        file: z.string(),
-        line: z.number(),
-        text: z.string(),
-      })),
-      unreadableSubtrees: z.array(z.string()),
-      unreadableFiles: z.array(z.string()),
-    }),
-    callback: async ({ query, typePattern, tier }: { query?: string; typePattern?: string; tier?: string }) => {
-      const mode: "name" | "type-pattern" = typePattern ? "type-pattern" : "name";
-      const actualQuery = (typePattern ?? query ?? "").trim();
-      if (actualQuery.length === 0) {
-        throw new ToolInvocationError({
-          message: "Provide either `query` or `typePattern`.",
-          classification: "invalid-input",
-          diagnostics: [
-            {
-              severity: "error",
-              message: "Provide either `query` or `typePattern`.",
-              code: "invalid-input",
-              nextAction:
-                "Pass `query` for a substring search across module / definition names, or `typePattern` for a token-pattern match against type signatures (e.g. `Nat → _ → Nat`).",
-            },
-          ],
-          data: { query: actualQuery, tier },
-        });
-      }
-      const requestedSearchRoot = tier
-        ? resolveFileWithinRoot(repoRoot, join("agda", tier))
-        : resolveFileWithinRoot(repoRoot, "agda");
-      if (!existsSync(requestedSearchRoot)) {
-        throw missingPathToolError("directory", requestedSearchRoot);
-      }
-      const searchRoot = resolveExistingPathWithinRoot(repoRoot, requestedSearchRoot);
-      // Ensure version detection has run so isAgdaSourceFile() filters by the
-      // actually-installed Agda's supported extensions rather than being
-      // permissive (all extensions) when this tool is invoked first.
-      if (!session.getAgdaVersion()) {
-        try { await session.query.showVersion(); } catch (err) {
-          logger.trace("agda_search_definitions: version detection best-effort failed", { err });
-        }
-      }
-      // Hard cap on the matches array so a pathological query
-      // (`pattern: " "`, etc.) on a huge repo can't OOM the server.
-      // The visible cap (slice(0, 50)) is much lower; this is the
-      // memory-safety backstop. 5000 matches × ~200 bytes/entry ≈ 1MB.
-      const MAX_RAW_MATCHES = 5000;
-      const matches: Array<{ file: string; line: number; text: string }> = [];
-      const unreadableDirs: string[] = [];
-      const unreadableFiles: string[] = [];
-      let truncatedAtCap = false;
-      function searchDir(dir: string, requestedDir: string, displayPrefix: string): void {
-        // Short-circuit the entire walk once we hit the cap. Without
-        // this check, a broad query on a huge repo continues recursing
-        // into sibling subtrees and reading more files even after
-        // `matches` is already full — pure I/O / CPU waste because
-        // the inner loop's cap check would just `return` after each
-        // file's first line. Bail at the directory boundary so an
-        // entire skipped subtree stops at one stat instead of N reads.
-        if (truncatedAtCap) return;
-        let entries: import("node:fs").Dirent[];
-        try {
-          entries = readdirSync(dir, { withFileTypes: true });
-        } catch (err) {
-          // Same graceful-walk contract as `agda_list_modules`: one
-          // unreadable subtree must not crash the search.
-          unreadableDirs.push(displayPrefix);
-          logger.trace("agda_search_definitions: readdir failed, skipping subtree", { dir, err });
-          return;
-        }
-        for (const entry of entries) {
-          if (truncatedAtCap) return;
-          const nextRequestedPath = resolve(requestedDir, entry.name);
-          const nextDisplayPath = join(displayPrefix, entry.name);
-          if (entry.isDirectory()) {
-            searchDir(resolve(dir, entry.name), nextRequestedPath, nextDisplayPath);
-          }
-          else if (isAgdaSourceFile(entry.name, session.getAgdaVersion() ?? undefined)) {
-            const filePath = resolveExistingChildWithinRoot(repoRoot, nextRequestedPath);
-            if (!filePath) {
-              continue;
-            }
-            let fileLines: string[];
-            try {
-              fileLines = readFileSync(filePath, "utf-8").split("\n");
-            } catch (err) {
-              // A file that readdir saw but readFile can't open
-              // (permission change mid-walk, truncated mid-read) is
-              // surfaced as a single "unreadable file" entry rather
-              // than dropped silently or crashing the tool.
-              unreadableFiles.push(nextDisplayPath);
-              logger.trace("agda_search_definitions: readFile failed", { file: filePath, err });
-              continue;
-            }
-            for (let i = 0; i < fileLines.length; i++) {
-              if (matches.length >= MAX_RAW_MATCHES) {
-                truncatedAtCap = true;
-                return;
-              }
-              const line = fileLines[i];
-              if (mode === "name") {
-                if (!line.includes(actualQuery)) continue;
-                matches.push({ file: nextDisplayPath, line: i + 1, text: line.trim() });
-                continue;
-              }
-              if (!line.includes(":")) continue;
-              const typePart = line.split(":").slice(1).join(":").trim();
-              if (!matchesTypePattern(typePart, actualQuery)) continue;
-              matches.push({ file: nextDisplayPath, line: i + 1, text: line.trim() });
-            }
-          }
-        }
-      }
-      searchDir(
-        searchRoot,
-        requestedSearchRoot,
-        relativeToRequestedRoot(repoRoot, requestedSearchRoot),
-      );
-      let output: string;
-      const capped = matches.slice(0, 50);
-      if (matches.length === 0) {
-        output = mode === "name"
-          ? `No matches for "${actualQuery}" in ${tier ?? "agda/"}`
-          : `No type-pattern matches for "${actualQuery}" in ${tier ?? "agda/"}`;
-      } else {
-        output = `## Search (${mode}): "${actualQuery}" (${matches.length} matches${matches.length > 50 ? ", showing first 50" : ""})\n\n`;
-        for (const m of capped) output += `- **${m.file}:${m.line}** \`${m.text}\`\n`;
-      }
-      if (unreadableDirs.length > 0 || unreadableFiles.length > 0) {
-        output += `\n_Skipped ${unreadableDirs.length} unreadable subtree(s)`;
-        if (unreadableFiles.length > 0) {
-          output += ` and ${unreadableFiles.length} unreadable file(s)`;
-        }
-        output += `; check file permissions or broken symlinks._`;
-      }
-      if (truncatedAtCap) {
-        output += `\n_Truncated at ${MAX_RAW_MATCHES} raw matches; refine the query to see the rest._`;
-      }
-      return {
-        text: output,
-        data: {
-          mode,
-          query: actualQuery,
-          tier: tier ?? null,
-          matchCount: matches.length,
-          shown: capped.length,
-          truncated: truncatedAtCap,
-          matches: capped,
-          unreadableSubtrees: unreadableDirs,
-          unreadableFiles,
-        },
-      };
-    },
-  });
+  registerReadModule(server, session, repoRoot);
+  registerListModules(server, session, repoRoot);
+  registerCheckPostulates(server, session, repoRoot);
+  registerSearchDefinitions(server, session, repoRoot);
 }
